+++ packages/file-bundle/src/RekalogikaFileBundle.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Rekalogika\File\Bundle;
 
 use Rekalogika\File\Bundle\DependencyInjection\FilterPass;
+use Rekalogika\File\Bundle\DependencyInjection\RemoveUnusedPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -30,6 +31,9 @@ final class RekalogikaFileBundle extends Bundle
     {
         parent::build($container);
 
-        $container->addCompilerPass(new FilterPass());
+        $container
+            ->addCompilerPass(new FilterPass())
+            ->addCompilerPass(new RemoveUnusedPass())
+        ;
     }
 }